Desenvolvimento: Criacão da pasta services para o processamento dos modos. Ajuste das dificuldades para o modo recreativo.
