+++ frontend/main.html
@@ -124,10 +124,10 @@
                     </p>
                     <div class="buttons-block-s">
                         <div class="buttons">
-                            <button class="button" id="btn-criativo">Modo Criativo</button>
+                            <button class="button" id="btn-criativo">Criativo</button>
                         </div>
                         <div class="buttons">
-                            <button class="button" id="btn-recreativo">Modo Recreativo</button>
+                            <button class="button" id="btn-recreativo">Recreativo</button>
                         </div>
                     </div>
                 </div>
